+++ packages/logging/test/unit/transports/axiom-js.test.ts
@@ -1,28 +1,28 @@
 import { describe, beforeEach, it, expect, vi } from 'vitest';
 import { AxiomJSTransport, axiomClient } from '../../../src/transports/axiom-js';
 import { createLogEvent } from '../../lib/mock';
-import { Axiom, AxiomWithoutBatching } from '@axiomhq/js';
+import { AxiomClient, AxiomClientWithoutBatching } from '@axiomhq/js';
 import { LogLevel } from 'src/logger';
 
 describe('AxiomJSTransport', () => {
-  let mockAxiom: Axiom;
-  let mockAxiomWithoutBatching: AxiomWithoutBatching;
+  let mockAxiom: AxiomClient;
+  let mockAxiomClientWithoutBatching: AxiomClientWithoutBatching;
   let transport: AxiomJSTransport;
   const DATASET = 'test-dataset';
 
   beforeEach(() => {
-    mockAxiomWithoutBatching = Object.create(AxiomWithoutBatching.prototype, {
+    mockAxiomClientWithoutBatching = Object.create(AxiomClientWithoutBatching.prototype, {
       ingest: { value: vi.fn() },
       appendAxiomClient: { value: vi.fn() },
     });
-    mockAxiom = Object.create(Axiom.prototype, {
+    mockAxiom = Object.create(AxiomClient.prototype, {
       ingest: { value: vi.fn() },
       flush: { value: vi.fn().mockResolvedValue(undefined) },
       appendAxiomClient: { value: vi.fn() },
     });
   });
 
-  describe('with Axiom', () => {
+  describe('with AxiomClient', () => {
     beforeEach(() => {
       transport = new AxiomJSTransport({ axiom: mockAxiom, dataset: DATASET });
     });
@@ -69,9 +69,9 @@ describe('AxiomJSTransport', () => {
     });
   });
 
-  describe('with AxiomWithoutBatching', () => {
+  describe('with AxiomClientWithoutBatching', () => {
     beforeEach(() => {
-      transport = new AxiomJSTransport({ axiom: mockAxiomWithoutBatching, dataset: DATASET });
+      transport = new AxiomJSTransport({ axiom: mockAxiomClientWithoutBatching, dataset: DATASET });
     });
 
     it('should forward logs to axiom client ingest method', () => {
@@ -79,8 +79,8 @@ describe('AxiomJSTransport', () => {
 
       transport.log(logs);
 
-      expect(mockAxiomWithoutBatching.ingest).toHaveBeenCalledTimes(1);
-      expect(mockAxiomWithoutBatching.ingest).toHaveBeenCalledWith(DATASET, logs);
+      expect(mockAxiomClientWithoutBatching.ingest).toHaveBeenCalledTimes(1);
+      expect(mockAxiomClientWithoutBatching.ingest).toHaveBeenCalledWith(DATASET, logs);
     });
 
     it('should resolve promises when flush is called', async () => {